requirements.txt
# Web框架
Django>=5.0.0
djangorestframework>=3.14.0

# AI和机器学习
openai>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.1.0

# 数据库
pymongo>=4.6.0
redis>=5.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL
mysqlclient>=2.2.0      # MySQL

# 异步和并发
aiohttp>=3.9.0
celery>=5.3.0
redis>=5.0.0

# 数据解析和爬虫
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0

# 测试和开发
coverage>=7.4.0
pytest>=7.4.0
pytest-django>=4.7.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# 工具和实用程序
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0

# 其他依赖
django-cors-headers>=4.3.1
django-filter>=23.5
Pillow>=10.1.0  # 用于图像处理
black>=23.11.0  # 代码格式化
isort>=5.12.0  # import排序
flake8>=6.1.0  # 代码检查 